back/requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/
aniso8601==9.0.1
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
dnspython==2.5.0; python_version >= '3.8'
email-validator==2.1.0.post1
et-xmlfile==1.1.0; python_version >= '3.6'
faker==22.6.0; python_version >= '3.8'
flask==2.2.5; python_version >= '3.7'
flask-cors==4.0.0
flask-mongoengine==1.0.0
flask-oidc@ git+https://github.com/puiterwijk/flask-oidc.git@f9e4367098f269160edca4c4601b98372aae8c65
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flask-uploads==0.2.1
flask-wtf==1.2.1; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0; python_version >= '3.5'
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.20.2; python_version >= '3.8'
mongoengine==0.27.0; python_version >= '3.7'
numpy==1.26.3; python_version < '3.11'
oauth2client==4.1.3
openpyxl==3.1.2; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandas==2.2.0; python_version >= '3.9'
pillow==10.2.0
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyjwt==2.8.0; python_version >= '3.7'
pymongo==4.6.1; python_version >= '3.7'
pymysql==1.1.0; python_version >= '3.7'
pyparsing==3.1.1; python_version >= '3.1'
pypng==0.20220715.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
qrcode[pil]==7.4.2; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
schedule==1.2.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.25; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.4; python_version >= '2'
urllib3==2.2.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
wtforms[email]==3.1.2; python_version >= '3.8'
